implement payment system and refactor project structure

+++ app/schemas/order.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from pydantic import BaseModel
 from app.schemas.movie import MovieResponse
 from app.models.orders import OrderStatusEnum
@@ -7,7 +8,7 @@ from app.models.orders import OrderStatusEnum
 class OrderItemResponse(BaseModel):
     id: int
     movie: MovieResponse
-    price_at_order: float
+    price_at_order: Decimal
 
     model_config = {"from_attributes": True}
